requirements.txt
asgiref==3.3.1
attrs==20.3.0
autobahn==20.7.1
Automat==20.2.0
certifi==2020.11.8
cffi==1.14.4
channels==3.0.2
chardet==3.0.4
configparser==5.0.1
constantly==15.1.0
cryptography==3.2.1
daphne==3.0.1
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.4
django-autoslug==1.9.8
django-ckeditor==6.0.0
django-cors-headers==3.5.0
django-debug-toolbar==3.2
django-extensions==3.1.0
django-import-export==2.4.0
django-js-asset==1.2.2
django-minio-storage==0.3.8
django-redis-cache==3.0.0
django-user-agents==0.4.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
jdcal==1.4.1
jwt==1.1.0
MarkupPy==1.14
minio==6.0.2
odfpy==1.4.1
openpyxl==3.0.5
Pillow==8.0.1
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==20.0.0
python-dateutil==2.8.1
python-magic==0.4.18
pytz==2020.4
PyYAML==5.3.1
redis==3.5.3
requests==2.25.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
tablib==2.0.0
Twisted==20.3.0
txaio==20.4.1
ua-parser==0.10.0
urllib3==1.26.2
user-agents==2.2.0
wincertstore==0.2
xlrd==1.2.0
xlwt==1.3.0
zope.interface==5.2.0
